Even better poisson disk
This one preserves its properties when combined with rotated versions of itself, reducing banding.

Its predecessor was overfitted as I'd got a parameter set far too high, which meant that each point was too close to being the most isolated one available when it was inserted instead of being random.
This meant that there was a lot of clustering if the radii were plotted, and in particular, the fifth point was nearly exactly at the centre, which made it invariant under rotation, causing a really pronounced band until there were loads of samples to outweigh it.

The new disk looks ridiculously much better at low sample counts, and still significantly better at higher ones.

It was generated with the same tool, but more settings were left at their default values, and it was in rotated disk mode with rotations as repetitions set to two.
As we really care about the distribution of radii, we don't want a higher value or it'd be too scared to put points close to the centre of the disk and too keen to put them near the outside.