Initial PCSS implementation
Requires upstream VSG changes to make the view dependent data descriptor set expose the shadow map as separate samplers and a texture instead of a single combined image sampler.
This is so the shadow map can be used both to access depth values and use hardware PCF via different samplers.

The blocker search is fairly standard and fairly straightforward.
Hopefully the code's fairly self-explanatory if you've got a vague idea of what PCSS is.

The penumbra radius calculation is a little less standard and less straightforward.
It prioritises correctness in the face of obstacles like light-space perspective shadow mapping or lights so large that the small angle approximation doesn't hold.
That means converting the average blocker back into eye space, which should definitely be Euclidean, using the inverse of the light-space matrix, computing the vector distance from the eye-space position, and actually calling tan on the angle.

This means that without any extra data passed from the CPU, we get a physically plausible penumbra size that's consistent between different shadow maps of different resolutions and scales.

On the other hand, larger penumbras make the pre-existing issue where shadow map boundaries are prominent even more visible.