Multi-cascade blocker search
Otherwise, the penumbra used to sample the tightest shadow map will be calculated without awareness of blockers just outside its bounds.

Also retune some parameters.

One metre is a more reliable blocker search radius now the above problems are fixed.
If it's too small, then the outer edges of penumbras could be lost.
If it's bigger, more samples are required to actually notice blockers, so samples ending up trustworthy helps.

The viable sample ratio seems mainly to optimise the cases where it's not safe to, so while it's not been removed yet, it's been set to one.