Spot light shadows
In principle, this should be everything that needs to change in this repo to make everything work as well as it does for directional lights.
Any further improvements will either happen in the main VSG repo, or be equally applicable to directional lights as the implementation is so similar.

In fact, the implementation for everything except PCSS is literally identical for spot and directional lights, so could be folded into the same function if that's preferred.

There's one just extra thing that might merit a spot-light-specific follow-up: shadowing is calculated before the spot light intensity, but it should be cheaper to do it the other way around, and skip shadow work outside the spotlight's cone of effect.