Fix alpha test for standard PBR shader
It looks like the existing line was copied and pasted from the Phong shader, but the variables don't match. Phong's `material`'s equivalent is called `pbr` in the PBR shader, and while `diffuseColor` is a real variable in the PBR shader, it's not been initialised when this line runs, so `baseColor` should be used instead.