vsgcameras demo of viewport partially off screen
The overlay views are partially off the screen. The initial rendering is good, but the model snaps to the scissors of the overlays when the window is resized. This is because the viewport gets set to the scissor in the viewportstate. It was expected that the viewport and the scissor remain independent and should be resized individually.